Modernize project infrastructure and switching to hatch (#108)
* Switch to src project structure

* Prevent CI being run twice on PR

Having CI jobs run on PR and push without any further restrictions runs
the CI jobs twice on a new PR, with this setting the CI is only run on
pushes to main.

* Ignore QbFields warning in sphinx

Bug that appeears with aiida 2.6 (independent of this PR), it also goes
away when building the docs twice. For now I ignore it.

* Remove outdated install_aiida_github.sh

It is nowhere used and I am not sure in which context it should be used.
Installing aiida-core seems trivial. What would be useful is something
for creating new profiles in the tests.

* Split pre-commit hook fmt into lint and format

* Change ruff settings single quotes -> double quotes

double quotes are default so nothing has to be changed

* Apply the formatting to double quotes on template repo

* Move docs to an optional dependency for readthedocs

Because readthedocs can only use extra dependencies from the
pyproject.toml, we move the deps of docs into one, so we can reference
it in the readthedocs.yaml

* Updates in the update-aiida-diff.sh script

- Remove `git init` in update aiida-diff script to base on current version
- Remove `hatch fmt` at the end since it is done by hook properly
* Fix formatting in the cookiecutter hook

Now `hatch fmt` is executed twice, running it one time did not fix all
isues. Needs to be investigated, if a reviewer has some ideas.

* Add script update-template-formatting.sh that can be used to fix formatting

To apply the formatting on the cookiecutter template the script initiates
a cookiecutter repo instance and applies the formatting on this instance
the changes are diffed and copied back to the template.

* Update update-aiida-diff.sh

Since the cookiecutter hooks is now also performing the formatting,
the pre-commit is not necessary. The git init has been removed since
the repository already exists.

---------

Co-authored-by: Rico Haeuselmann <r.haeuselmann@gmx.ch>